MDEV-25129 postfix for windows
C:\projects\server\sql\sql_show.cc(7913): error C2220: warning treated as error - no 'object' file generated [C:\projects\server\win_build\sql\sql.vcxproj]
C:\projects\server\sql\sql_show.cc(7913): warning C4267: 'initializing': conversion from 'size_t' to 'uint', possible loss of data [C:\projects\server\win_build\sql\sql.vcxproj]

caused by 768c51880a5aa6d25d4c0fe7de7a88561ff46422

+++ sql/sql_show.cc
@@ -7572,7 +7572,7 @@ int fill_variables(THD *thd, TABLE_LIST *tables, COND *cond)
 int add_symbol_to_table(const char* name, TABLE* table){
   DBUG_ENTER("add_symbol_to_table");
 
-  uint length= strlen(name);
+  size_t length= strlen(name);
 
   // If you've added a new SQL keyword longer than KEYWORD_SIZE,
   // please increase the defined max length
